Make floattype depend on the total number of bits, not just fractional bits
This ensures that conversion is lossless even for the largest representable values.

+++ test/ufixed.jl
@@ -112,7 +112,7 @@ x = UFixed8(0b01010001, 0)
 @test -x == 0xafuf8
 
 @test isa(float(one(UFixed{UInt8,7})),   Float32)
-@test isa(float(one(UFixed{UInt32,18})), Float32)
+@test isa(float(one(UFixed{UInt32,18})), Float64)
 @test isa(float(one(UFixed{UInt32,25})), Float64)
 
 for T in (FixedPointNumbers.UF..., UF2...)
